last version

+++ src/app/admin/users/listuser/listuser.ts
@@ -5,10 +5,13 @@ import { OnInit } from '@angular/core';
 import { AuthService } from '../../../services/authentification';
 import { FormsModule } from '@angular/forms';
 import { MatIconModule } from '@angular/material/icon';
+import { ConfirmationPopupComponent } from '../../../shared/confirmation-popup/confirmation-popup/confirmation-popup';
+import { MatDialog } from '@angular/material/dialog';
+import { DatePipe } from '@angular/common';
 
 @Component({
   selector: 'app-listuser',
-  imports: [RouterModule, FormsModule, MatIconModule],
+  imports: [RouterModule, FormsModule, MatIconModule, DatePipe],
   templateUrl: './listuser.html',
   styleUrl: './listuser.css',
 })
@@ -16,8 +19,7 @@ export class Listuser implements OnInit {
   Users: any[] = [];
   filteredUsers: any[] = [];
   searchText: string = '';
-
-  constructor(private authService: AuthService, private userService: UserService, private router: Router) { }
+  constructor(private dialog: MatDialog, private authService: AuthService, private userService: UserService, private router: Router) { }
   ngOnInit(): void {
     if (!this.authService.isLoggedIn()) {
       this.router.navigate(['/login']);
@@ -30,7 +32,9 @@ export class Listuser implements OnInit {
     this.filteredUsers = this.Users.filter(user =>
       (user.email?.toLowerCase().includes(value)) ||
       (user.firstname?.toLowerCase().includes(value)) ||
-      (user.lastname?.toLowerCase().includes(value))
+      (user.lastname?.toLowerCase().includes(value)) ||
+      (user.role?.toLowerCase().includes(value)) ||
+      (user.departement?.toLowerCase().includes(value))
     );
   }
   getUserList(): void {
@@ -41,13 +45,24 @@ export class Listuser implements OnInit {
       },
     });
   }
-
   deleteUser(userId: number): void {
-    this.userService.deleteUser(userId).subscribe({
-      next: () => {
-        this.getUserList();
-      },
+    const dialogRef = this.dialog.open(ConfirmationPopupComponent, {
+      width: '600px',
+      data: {
+        title: 'Suppression',
+        message: 'Voulez-vous vraiment supprimer cet utilisateur ?',
+        confirmText: 'Supprimer',
+        cancelText: 'Annuler'
+      }
+    });
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.userService.deleteUser(userId).subscribe({
+          next: () => {
+            this.getUserList();
+          }
+        });
+      }
     });
   }
-
 }
